requirements.txt
# Web Framework and API
streamlit==1.29.0
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# Database
SQLAlchemy==2.0.23
aiosqlite==0.19.0

# Face Recognition and Image Processing
face-recognition==1.3.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
Pillow==10.1.0

# Data Processing and Visualization
pandas==2.1.4
plotly==5.18.0

# HTTP Client
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pydantic==2.5.2
